Plus-minus-statement
Given an array of integers, calculate the fractions of its elements that are positive, negative, and are zeros. Print the decimal value of each fraction on a new line.

For example, given the array arr= [1,1,0,-1,-2] there are 5 elements, two positive, two negative and one zero. Their ratios would be 2/5 = 0.400000, 2/5 = 0.400000 and 1/5 = 0.200000. It should be printed as 
0.400000
0.400000
0.200000

Function Description:

Complete the plusMinus function in the editor below. It should print out the ratio of positive, negative and zero items in the array, each on a separate line rounded to six decimals.

plusMinus has the following parameter(s):
arr: an array of integers


https://www.hackerrank.com/contests/target-codevita-4/challenges/plus-minus